arange returns type equivalent to intc

+++ numba/targets/arrayobj.py
@@ -3765,7 +3765,7 @@ def _arange_dtype(*args):
     elif any(isinstance(a, types.Float) for a in bounds):
         dtype = types.float64
     else:
-        dtype = types.intp
+        dtype = types.intc
 
     return dtype
